[refactor] change compoent file name
