+++ package.json
@@ -10,6 +10,7 @@
     "@testing-library/react": "^10.0.1",
     "@testing-library/user-event": "^10.0.0",
     "dotenv": "^8.2.0",
+    "firebase": "^7.13.1",
     "minimist": "^1.2.5",
     "node-sass": "^4.13.0",
     "react": "^16.12.0",
