+++ src/components/collection-item/collection-item.compoent.js
