+++ src/pages/homepage/homepage.component.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import './homepage.scss'
-import Directory from '../../components/directory'
+import Directory from '../../components/directory/directory.component'
 
 const HomePage = props => {
   console.log(props)
