+++ src/pages/shop/shop.component.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 
 import collectionsFaker from '../../fakers/collections-faker'
-import CollectionPreview from '../../components/collection-preview'
+import CollectionPreview from '../../components/collection-preview/collection-preview.component'
 
 const ShopPage = props => {
   const [collections, setCollections] = useState(collectionsFaker)
